requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.6.0
django-filter==24.3
django-haystack==3.3.0
django-meta==2.3.0
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0
gunicorn==23.0.0
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
sentry-sdk==2.19.2
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Whoosh==2.7.4
psutil==6.0.0
paho-mqtt==1.6.1
